requirements.txt
# MCP SDK
mcp>=1.2.0

# Gmail API
google-auth>=2.43.0
google-auth-oauthlib>=1.2.0
google-auth-httplib2>=0.2.0
google-api-python-client>=2.149.0

# Configuration
python-dotenv>=1.0.0

# Utilities
pytz>=2024.1
python-dateutil>=2.9.0

# Testing
pytest>=8.0.0
pytest-asyncio>=0.24.0
